LPS-63359 com.liferay.portal.search.* packages
https://issues.liferay.com/browse/LPS-63359
