+++ modules/portal-search/portal-search-facet/src/main/java/com/liferay/portal/search/facet/BaseJSPSearchFacet.java
@@ -1,0 +1,95 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.facet;
+
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.io.IOException;
+
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * @author Eudaldo Alonso
+ */
+public abstract class BaseJSPSearchFacet extends BaseSearchFacet {
+
+	public abstract String getConfigurationJspPath();
+
+	public abstract String getDisplayJspPath();
+
+	@Override
+	public void includeConfiguration(
+			HttpServletRequest request, HttpServletResponse response)
+		throws IOException {
+
+		if (Validator.isNull(getConfigurationJspPath())) {
+			return;
+		}
+
+		RequestDispatcher requestDispatcher =
+			_servletContext.getRequestDispatcher(getConfigurationJspPath());
+
+		try {
+			requestDispatcher.include(request, response);
+		}
+		catch (ServletException se) {
+			_log.error("Unable to include JSP " + getDisplayJspPath(), se);
+
+			throw new IOException(
+				"Unable to include " + getDisplayJspPath(), se);
+		}
+	}
+
+	@Override
+	public void includeView(
+			HttpServletRequest request, HttpServletResponse response)
+		throws IOException {
+
+		if (Validator.isNull(getDisplayJspPath())) {
+			return;
+		}
+
+		RequestDispatcher requestDispatcher =
+			_servletContext.getRequestDispatcher(getDisplayJspPath());
+
+		try {
+			requestDispatcher.include(request, response);
+		}
+		catch (ServletException se) {
+			if (_log.isErrorEnabled()) {
+				_log.error("Unable to include JSP", se);
+			}
+
+			throw new IOException(
+				"Unable to include " + getDisplayJspPath(), se);
+		}
+	}
+
+	public void setServletContext(ServletContext servletContext) {
+		_servletContext = servletContext;
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		BaseJSPSearchFacet.class);
+
+	private ServletContext _servletContext;
+
+}
