+++ modules/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/constants/SearchPortletParameterNames.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.web.constants;
+
+/**
+ * @author André de Oliveira
+ */
+public class SearchPortletParameterNames {
+
+	public static final String FORMAT = "format";
+
+	public static final String GROUP_ID = "groupId";
+
+	public static final String KEYWORDS = "keywords";
+
+	public static final String SCOPE = "scope";
+
+}
