+++ modules/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/portlet/SearchViewPortletProvider.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.web.portlet;
+
+import com.liferay.admin.kernel.util.PortalSearchApplicationType;
+import com.liferay.portal.kernel.portlet.BasePortletProvider;
+import com.liferay.portal.kernel.portlet.ViewPortletProvider;
+import com.liferay.portal.search.web.constants.SearchPortletKeys;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Eudaldo Alonso
+ */
+@Component(
+	immediate = true,
+	property = {
+		"model.class.name=" + PortalSearchApplicationType.Search.CLASS_NAME
+	},
+	service = ViewPortletProvider.class
+)
+public class SearchViewPortletProvider
+	extends BasePortletProvider implements ViewPortletProvider {
+
+	@Override
+	public String getPortletId() {
+		return SearchPortletKeys.SEARCH;
+	}
+
+}
