--- modules/portal-search/portal-search-web/src/main/java/com/liferay/search/web/facet/FolderSearchFacet.java
@@ -1,118 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.search.web.facet;
-
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.facet.MultiValueFacet;
-import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.search.facet.BaseJSPSearchFacet;
-import com.liferay.search.facet.SearchFacet;
-
-import javax.portlet.ActionRequest;
-
-import javax.servlet.ServletContext;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Eudaldo Alonso
- */
-@Component(immediate = true, service = SearchFacet.class)
-public class FolderSearchFacet extends BaseJSPSearchFacet {
-
-	@Override
-	public String getConfigurationJspPath() {
-		return "/facets/configuration/folders.jsp";
-	}
-
-	@Override
-	public FacetConfiguration getDefaultConfiguration(long companyId) {
-		FacetConfiguration facetConfiguration = new FacetConfiguration();
-
-		facetConfiguration.setClassName(getFacetClassName());
-
-		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
-
-		jsonObject.put("frequencyThreshold", 1);
-		jsonObject.put("maxTerms", 10);
-		jsonObject.put("showAssetCount", true);
-
-		facetConfiguration.setDataJSONObject(jsonObject);
-
-		facetConfiguration.setFieldName(getFieldName());
-		facetConfiguration.setLabel(getLabel());
-		facetConfiguration.setOrder(getOrder());
-		facetConfiguration.setStatic(false);
-		facetConfiguration.setWeight(1.2);
-
-		return facetConfiguration;
-	}
-
-	@Override
-	public String getDisplayJspPath() {
-		return "/facets/view/folders.jsp";
-	}
-
-	@Override
-	public String getFacetClassName() {
-		return MultiValueFacet.class.getName();
-	}
-
-	@Override
-	public String getFieldName() {
-		return Field.FOLDER_ID;
-	}
-
-	@Override
-	public JSONObject getJSONData(ActionRequest actionRequest) {
-		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
-
-		int frequencyThreshold = ParamUtil.getInteger(
-			actionRequest, getClassName() + "frequencyThreshold", 1);
-		int maxTerms = ParamUtil.getInteger(
-			actionRequest, getClassName() + "maxTerms", 10);
-		boolean showAssetCount = ParamUtil.getBoolean(
-			actionRequest, getClassName() + "showAssetCount", true);
-
-		jsonObject.put("frequencyThreshold", frequencyThreshold);
-		jsonObject.put("maxTerms", maxTerms);
-		jsonObject.put("showAssetCount", showAssetCount);
-
-		return jsonObject;
-	}
-
-	@Override
-	public String getLabel() {
-		return "any-folder";
-	}
-
-	@Override
-	public String getTitle() {
-		return "folder";
-	}
-
-	@Override
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.search.web)", unbind = "-"
-	)
-	public void setServletContext(ServletContext servletContext) {
-		super.setServletContext(servletContext);
-	}
-
-}
