--- modules/portal-search/portal-search-web/src/main/java/com/liferay/search/web/upgrade/SearchWebUpgrade.java
@@ -1,40 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.search.web.upgrade;
-
-import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
-import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.portal.upgrade.util.UpgradePortletId;
-import com.liferay.search.web.upgrade.v1_0_0.UpgradePortletPreferences;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Eudaldo Alonso
- */
-@Component(immediate = true, service = UpgradeStepRegistrator.class)
-public class SearchWebUpgrade implements UpgradeStepRegistrator {
-
-	@Override
-	public void register(Registry registry) {
-		registry.register(
-			"com.liferay.search.web", "0.0.0", "1.0.0", new DummyUpgradeStep());
-
-		registry.register(
-			"com.liferay.search.web", "0.0.1", "1.0.0", new UpgradePortletId(),
-			new UpgradePortletPreferences());
-	}
-
-}
