+++ modules/portal-search/portal-search-web/src/main/resources/META-INF/resources/init.jsp
@@ -83,14 +83,14 @@ page import="com.liferay.portal.kernel.util.StringUtil" %><%@
 page import="com.liferay.portal.kernel.util.Tuple" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
 page import="com.liferay.portal.kernel.xml.Element" %><%@
+page import="com.liferay.portal.search.facet.SearchFacet" %><%@
+page import="com.liferay.portal.search.facet.util.SearchFacetTracker" %><%@
+page import="com.liferay.portal.search.facet.util.comparator.SearchFacetComparator" %><%@
+page import="com.liferay.portal.search.web.constants.SearchPortletKeys" %><%@
+page import="com.liferay.portal.search.web.constants.SearchPortletParameterNames" %><%@
+page import="com.liferay.portal.search.web.display.context.SearchDisplayContext" %><%@
+page import="com.liferay.portal.search.web.facet.AssetEntriesSearchFacet" %><%@
 page import="com.liferay.portlet.asset.service.permission.AssetCategoryPermission" %><%@
-page import="com.liferay.search.facet.SearchFacet" %><%@
-page import="com.liferay.search.facet.util.SearchFacetTracker" %><%@
-page import="com.liferay.search.facet.util.comparator.SearchFacetComparator" %><%@
-page import="com.liferay.search.web.constants.SearchPortletKeys" %><%@
-page import="com.liferay.search.web.constants.SearchPortletParameterNames" %><%@
-page import="com.liferay.search.web.display.context.SearchDisplayContext" %><%@
-page import="com.liferay.search.web.facet.AssetEntriesSearchFacet" %><%@
 page import="com.liferay.taglib.aui.AUIUtil" %><%@
 page import="com.liferay.taglib.servlet.PipingServletResponse" %>
