+++ modules/portal-search/portal-search-web/src/main/resources/META-INF/resources/main_search.jspf
@@ -103,7 +103,7 @@ Hits hits = searchDisplayContext.getHits();
 					assetRenderer = assetRendererFactory.getAssetRenderer(classPK);
 				}
 
-				String viewURL = com.liferay.search.web.util.SearchUtil.getSearchResultViewURL(renderRequest, renderResponse, className, classPK, searchDisplayContext.isViewInContext(), currentURL);
+				String viewURL = com.liferay.portal.search.web.util.SearchUtil.getSearchResultViewURL(renderRequest, renderResponse, className, classPK, searchDisplayContext.isViewInContext(), currentURL);
 
 				Indexer indexer = IndexerRegistryUtil.getIndexer(className);
